Extract Cashu wallet storage
