+++ service-worker.js
@@ -385,6 +385,7 @@ const APP_SHELL = [
   '/js/supplement-warnings.js',
   '/js/food-contaminants.js',
   '/js/cashu-wallet.js',
+  '/js/cashu-wallet-store.js',
   '/js/nostr-discovery.js',
   '/js/touch-tooltip-runtime.js',
   '/js/touch-tooltip.js',
